% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulsab-cv-package.R
\docType{package}
\name{paulsab.cv-package}
\alias{paulsab.cv}
\alias{paulsab.cv-package}
\title{paulsab.cv: A package to generate my CV}
\description{
Builds a CV using the brilliant vitae package by Mitchell O'Hara-Wild.
}
\author{
\strong{Maintainer}: Paul Sabournkov \email{paulsab@gmail.com}

}
\keyword{internal}
